21:53

+++ include.h
@@ -32,6 +32,6 @@ char		*ft_clearrem_gnl(char *rem_line);
 int	ft_word_count(char const *s, char c);
 //fdf functions
 void get_map_info( char *f_name, t_info *info);
-void ft_bresenham_draw_line(int x1, int y1, int x2, int y2, t_info *info);
-
+//void ft_bresenham_draw_line(int x1, int y1, int x2, int y2, t_info *info);
+void ft_bresenham_draw_line(float x1, float y1, float x2, float y2, t_info *info);
 #endif
